Split_InputField
