+++ main/nextpatchnumber.txt
@@ -1,1 +1,1 @@
-24
+25
